+++ src/features/mypage/views/deliveryaddress/DeliveryAddressCreateView.vue
@@ -1,6 +1,6 @@
 <script setup>
 
-import DeliveryAddressForm from '@/features/mypage/components/DeliveryAddressForm.vue';
+import DeliveryAddressForm from '@/features/mypage/components/form/delivery/DeliveryAddressForm.vue';
 </script>
 
 <template>
