+++ src/features/mypage/views/like/LikeListView.vue
@@ -3,8 +3,8 @@ import { ref, computed, onMounted, watch } from 'vue';
 import PaginationBar from '@/components/common/PaginationBar.vue';
 import { useRoute, useRouter } from 'vue-router';
 import ItemCard from '@/components/common/ItemCard.vue';
-import FilterHeader from '@/features/mypage/components/FilterHeader.vue';
-import MyPageHeader from '@/features/mypage/components/MyPageHeader.vue';
+import FilterHeader from '@/features/mypage/components/common/FilterHeader.vue';
+import MyPageHeader from '@/features/mypage/components/common/MyPageHeader.vue';
 
 import likes from '@/features/mypage/mock/likes.json';
 import packages from '@/features/package/mock/packages.json';
